backend/pyproject.toml
[project]
name = "mnist-backend"
version = "0.1.0"
description = "Backend API for MNIST digit recognition"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "tensorflow[and-cuda]>=2.18.0",
    "tensorflow-datasets>=4.9.0",
    "keras>=3.7.0",
    "numpy>=2.2.0",
    "pillow>=11.0.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
